# internal/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Better Lyrics API
  version: 1.0.0
  summary: Synchronized lyrics for Better Lyrics.
  description: |
    Fetches word and line synced lyrics, caches them, and serves them to the
    Better Lyrics extension.

    Query parameters with aliases (for example `s`, `song`, `songName`) are
    concatenated when more than one alias is sent, so send only one of each.

    Every response passes through a two-tier, per-IP rate limiter. The normal
    tier serves everything. When it is exhausted, the cached tier serves only
    cached lyrics. When both are exhausted the server returns `429` with
    `Retry-After: 1`. Rate limited responses carry `X-RateLimit-Limit`,
    `X-RateLimit-Remaining` and `X-RateLimit-Type`. A valid `X-API-Key` skips
    rate limiting and sends `X-RateLimit-Bypass: true` instead.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
...
servers:
  - url: https://api.betterlyrics.org
...
paths:
  /getLyrics:
    get:
      tags: [Lyrics]
      operationId: getLyrics
      summary: Get lyrics (default provider)
      description: |
        Returns TTML lyrics from the default provider. Send a song, an artist,
        or both. Cache lookups tolerate small duration differences. When the
        server enforces API keys, a cache miss needs a valid `X-API-Key`;
        cache hits never do.
      security:
        - {}
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/s'
        - $ref: '#/components/parameters/song'
        - $ref: '#/components/parameters/songName'
        - $ref: '#/components/parameters/a'
        - $ref: '#/components/parameters/artist'
        - $ref: '#/components/parameters/artistName'
        - $ref: '#/components/parameters/al'
        - $ref: '#/components/parameters/album'
        - $ref: '#/components/parameters/albumName'
        - $ref: '#/components/parameters/d'
        - $ref: '#/components/parameters/duration'
        - name: videoId
          in: query
          description: Video ID. Links the video to the song so downstream caches can be revalidated.
          schema:
            type: string
        - name: v
          in: query
          description: Alias of `videoId`.
          schema:
            type: string
      responses:
        '200':
          description: Lyrics found.
          headers: &lyricsHeaders
            X-Cache-Status:
              $ref: '#/components/headers/X-Cache-Status'
            X-RateLimit-Type:
              $ref: '#/components/headers/X-RateLimit-Type'
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Bypass:
              $ref: '#/components/headers/X-RateLimit-Bypass'
            X-Auth-Mode:
              $ref: '#/components/headers/X-Auth-Mode'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TTMLLyrics'
        '401':
          description: |
            The query is not cached and the server enforces API keys, but no
            valid `X-API-Key` was sent.
          headers: *lyricsHeaders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWithMessage'
        '404':
          description: No lyrics exist for this song. This result is cached too (`NEGATIVE_HIT`).
          headers: *lyricsHeaders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/MissingSongAndArtist'
        '429':
          description: |
            Rate limited. `text/plain` with `Retry-After: 1` when both tiers
            are exhausted. JSON with `Retry-After: 60` and `X-Cache-Status`
            when the cached tier is in use and the query is not cached.
          headers:
            Retry-After:
              $ref: '#/components/headers/Retry-After'
            X-RateLimit-Type:
              $ref: '#/components/headers/X-RateLimit-Type'
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-Cache-Status:
              description: Set on the JSON variant only.
              schema:
                $ref: '#/components/schemas/CacheStatus'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWithMessage'
            text/plain:
              schema:
                type: string
        '500':
          description: Upstream failure with no cached fallback.
          headers: *lyricsHeaders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: The server is in cache-only mode and the query is not cached.
          headers: *lyricsHeaders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /ttml/getLyrics:
    get:
      tags: [Lyrics]
      operationId: getTTMLLyrics
      summary: Get lyrics from the TTML provider
      description: |
        Syllable-level timing. Send a song, an artist, or both. Uses exact
        cache keys, with no duration tolerance. When the server enforces API
        keys, a cache miss needs a valid `X-API-Key`.
      security:
        - {}
        - apiKey: []
      parameters: &providerParameters
        - $ref: '#/components/parameters/s'
        - $ref: '#/components/parameters/song'
        - $ref: '#/components/parameters/songName'
        - $ref: '#/components/parameters/a'
        - $ref: '#/components/parameters/artist'
        - $ref: '#/components/parameters/artistName'
        - $ref: '#/components/parameters/al'
        - $ref: '#/components/parameters/album'
        - $ref: '#/components/parameters/albumName'
        - $ref: '#/components/parameters/d'
        - $ref: '#/components/parameters/duration'
      responses:
        '200':
          description: Lyrics found.
          headers: &ttmlProviderHeaders
            X-Cache-Status:
              $ref: '#/components/headers/X-Cache-Status'
            X-Provider:
              $ref: '#/components/headers/X-Provider'
            X-RateLimit-Type:
              $ref: '#/components/headers/X-RateLimit-Type'
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Bypass:
              $ref: '#/components/headers/X-RateLimit-Bypass'
            X-Auth-Mode:
              $ref: '#/components/headers/X-Auth-Mode'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderLyrics'
        '401':
          description: |
            The query is not cached and the server enforces API keys, but no
            valid `X-API-Key` was sent.
          headers: *ttmlProviderHeaders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderAPIKeyError'
        '404':
          description: No lyrics exist for this song from this provider.
          headers: *ttmlProviderHeaders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderNotFound'
        '422':
          $ref: '#/components/responses/MissingSongAndArtist'
        '429':
          $ref: '#/components/responses/ProviderRateLimited'
        '500':
          description: Upstream failure.
          headers: *ttmlProviderHeaders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderError'
        '503':
          description: The server is in cache-only mode and the query is not cached.
          headers: *ttmlProviderHeaders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderError'
  /kugou/getLyrics:
    get:
      tags: [Lyrics]
      operationId: getKugouLyrics
      summary: Get lyrics from the Kugou provider
      description: Line-level timing. Send a song, an artist, or both. Uses exact cache keys, with no duration tolerance.
      security:
        - {}
        - apiKey: []
      parameters: *providerParameters
      responses: &publicProviderResponses
        '200':
          description: Lyrics found.
          headers: &providerHeaders
            X-Cache-Status:
              $ref: '#/components/headers/X-Cache-Status'
            X-Provider:
              $ref: '#/components/headers/X-Provider'
            X-RateLimit-Type:
              $ref: '#/components/headers/X-RateLimit-Type'
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Bypass:
              $ref: '#/components/headers/X-RateLimit-Bypass'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderLyrics'
        '404':
          description: No lyrics exist for this song from this provider.
          headers: *providerHeaders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderNotFound'
        '422':
          $ref: '#/components/responses/MissingSongAndArtist'
        '429':
          $ref: '#/components/responses/ProviderRateLimited'
        '500':
          description: Upstream failure.
          headers: *providerHeaders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderError'
        '503':
          description: The server is in cache-only mode and the query is not cached.
          headers: *providerHeaders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderError'
  /qq/getLyrics:
    get:
      tags: [Lyrics]
      operationId: getQQLyrics
      summary: Get lyrics from the QQ provider
      description: Send a song, an artist, or both. Uses exact cache keys, with no duration tolerance.
      security:
        - {}
        - apiKey: []
      parameters: *providerParameters
      responses: *publicProviderResponses
...
  /openapi.json:
    get:
      tags: [Meta]
      operationId: getOpenAPI
      summary: This OpenAPI document
      responses:
        '200':
          description: OpenAPI 3.1 document.
          content:
            application/json:
              schema:
                type: object
                required: [openapi, info, paths]
        '429':
          $ref: '#/components/responses/RateLimited'
  /cache:
    get:
      tags: [Cache]
      operationId: getCacheDump
      summary: Removed full cache dump
      deprecated: true
      responses:
        '410':
          description: Endpoint removed. The body lists alternatives.
          headers: &cacheHeaders
            X-RateLimit-Type:
              $ref: '#/components/headers/X-RateLimit-Type'
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [error, message, alternatives]
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  alternatives:
                    type: object
                    additionalProperties:
                      type: string
        '429':
          $ref: '#/components/responses/RateLimited'
  /cache/help:
    get:
      tags: [Cache]
      operationId: getCacheHelp
      summary: Cache endpoint documentation
      responses:
        '200':
          description: Human-readable help.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [description, endpoints, cache_key_format, notes]
                properties:
                  description:
                    type: string
                  endpoints:
                    type: array
                    items:
                      type: object
                      additionalProperties: false
                      required: [path, method, auth, description]
                      properties:
                        path:
                          type: string
                        method:
                          type: string
                        auth:
                          type: string
                        description:
                          type: string
                  cache_key_format:
                    type: object
                    additionalProperties:
                      type: string
                  notes:
                    type: array
                    items:
                      type: string
        '429':
          $ref: '#/components/responses/RateLimited'
  /cache/lookup:
    get:
      tags: [Cache]
      operationId: cacheLookup
      summary: Check whether a song is cached
      description: Send a song, an artist, or both.
      security:
        - adminToken: []
      parameters: *providerParameters
      responses:
        '200':
          description: Lookup result, including the cache keys the query maps to.
          headers: *cacheHeaders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheLookup'
        '400':
          description: Song and artist both missing.
          headers: *cacheHeaders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Skips rate limiting. When the server enforces API keys, it is also
        required for cache misses on `/getLyrics`, `/ttml/getLyrics` and for
        `/revalidate`.
    adminToken:
      type: apiKey
      in: header
      name: Authorization
      description: The raw admin token, with no `Bearer` prefix.
  parameters:
    s:
      name: s
      in: query
      description: Song name.
      schema:
        type: string
      example: Viva la Vida
    song:
      name: song
      in: query
      description: Alias of `s`.
      schema:
        type: string
    songName:
      name: songName
      in: query
      description: Alias of `s`.
      schema:
        type: string
    a:
      name: a
      in: query
      description: Artist name.
      schema:
        type: string
      example: Coldplay
    artist:
      name: artist
      in: query
      description: Alias of `a`.
      schema:
        type: string
...
  schemas:
    CacheStatus:
      type: string
      enum: [HIT, MISS, STALE, NEGATIVE_HIT]
...